Added switch case option
Now there is two ways to code this. I cant get the first (else if) to work, only displays first if. The second option is operational, but only works if asking for one value (cant say "case 3||4". note: could do the cases but would have to have 2 cases for the ones with two options. That would leave the >8 people option up to solve.